Add indexing test with RamBuffer

+++ src/main/java/com/qwazr/search/bench/LuceneCommonIndex.java
@@ -41,8 +41,8 @@ import java.util.concurrent.ExecutorService;
  */
 abstract public class LuceneCommonIndex implements Closeable {
 
-	private final static int MAX_SSD_MERGE_THREADS =
-			Math.max(1, Math.min(4, Runtime.getRuntime().availableProcessors() / 2));
+	private final static int MAX_SSD_MERGE_THREADS = Math.max(1,
+			Math.min(4, Runtime.getRuntime().availableProcessors() / 2));
 
 	final Path indexDirectory;
 	final Directory dataDirectory;
@@ -50,14 +50,14 @@ abstract public class LuceneCommonIndex implements Closeable {
 	final IndexWriter indexWriter;
 
 	LuceneCommonIndex(final Path rootDirectory, final String schemaName, final String indexName,
-			final int ramBufferSize) throws IOException {
+			final double ramBufferSize) throws IOException {
 
 		final Path schemaDirectory = Files.createDirectory(rootDirectory.resolve(schemaName));
 		this.indexDirectory = Files.createDirectory(schemaDirectory.resolve(indexName));
 
 		this.dataDirectory = FSDirectory.open(indexDirectory.resolve("data"));
-		final IndexWriterConfig indexWriterConfig =
-				new IndexWriterConfig(new PerFieldAnalyzerWrapper(new StandardAnalyzer()));
+		final IndexWriterConfig indexWriterConfig = new IndexWriterConfig(
+				new PerFieldAnalyzerWrapper(new StandardAnalyzer()));
 		indexWriterConfig.setOpenMode(IndexWriterConfig.OpenMode.CREATE_OR_APPEND);
 		indexWriterConfig.setRAMBufferSizeMB(ramBufferSize);
 		final ConcurrentMergeScheduler mergeScheduler = new ConcurrentMergeScheduler();
@@ -66,8 +66,8 @@ abstract public class LuceneCommonIndex implements Closeable {
 		indexWriterConfig.setMergePolicy(new TieredMergePolicy());
 
 		// We use snapshots deletion policy
-		final SnapshotDeletionPolicy snapshotDeletionPolicy =
-				new SnapshotDeletionPolicy(indexWriterConfig.getIndexDeletionPolicy());
+		final SnapshotDeletionPolicy snapshotDeletionPolicy = new SnapshotDeletionPolicy(
+				indexWriterConfig.getIndexDeletionPolicy());
 		indexWriterConfig.setIndexDeletionPolicy(snapshotDeletionPolicy);
 
 		this.indexWriter = new IndexWriter(this.dataDirectory, indexWriterConfig);
