+++ src/main/java/com/qwazr/search/bench/LuceneNoTaxonomyIndex.java
@@ -34,7 +34,7 @@ public class LuceneNoTaxonomyIndex extends LuceneCommonIndex {
 	private final SearcherManager searcherManager;
 
 	public LuceneNoTaxonomyIndex(final Path rootDirectory, final String schemaName, final String indexName,
-			final ExecutorService executorService, final int ramBufferSize) throws IOException {
+			final ExecutorService executorService, final double ramBufferSize) throws IOException {
 		super(rootDirectory, schemaName, indexName, ramBufferSize);
 		searcherManager = new SearcherManager(this.indexWriter,
 				executorService == null ? new SearcherFactory() : new MultiThreadSearcherFactory(executorService));
